+++ src/bcanalysis.h
@@ -22,7 +22,8 @@ class bcanalysis{
 		
 		int importRawSignal(const char *,int,const char *,const char *);//Implemented
 		int processSignals(const char *,int);//Implemented
-		int outputData(const char *, unsigned short int,unsigned short int,unsigned short int[]);//Implemented
+//		int outputData(const char *, unsigned short int,unsigned short int,unsigned short int[]);//Implemented
+		int outputData(const char *, unsigned short int,unsigned short int,unsigned int[]);//Implemented
 		int importTagAlign(const char *,int);
 		int importBowtie(const char *,int);
 		int importBed(const char *,int);
